Move sentry interceptor to nucliadb_telemetry
