+++ nucliadb_utils/nucliadb_utils/grpc.py
@@ -25,8 +25,8 @@ from grpc import aio  # type: ignore
 from grpc import ChannelCredentials
 
 from nucliadb_telemetry.grpc import GRPCTelemetry
+from nucliadb_telemetry.grpc_sentry import SentryInterceptor
 from nucliadb_telemetry.utils import get_telemetry
-from nucliadb_utils.sentry_interceptor import SentryInterceptor
 
 logger = logging.getLogger(__name__)
